requirements.txt
Flask>=2.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
cohere>=2.0
PyMuPDF>=1.21   # provides `import fitz`
gunicorn>=20.0.0